+++ tests/test_build_mask.py
@@ -21,6 +21,7 @@
 """
 import os
 import unittest
+import warnings
 
 import hydra
 import numpy as np
@@ -57,6 +58,10 @@ suffix_dict = {"PNG": ".png", "GTiff": ".tif", "JPEG": ".jpg"}
 
 class Test_TestBuildMask(unittest.TestCase):
     def setUp(self):
+        warnings.simplefilter("ignore", category=ImportWarning)
+        warnings.simplefilter("ignore", category=DeprecationWarning)
+        warnings.simplefilter("ignore", category=FutureWarning)
+        warnings.simplefilter("ignore", category=UserWarning)
         self.output_dir = create_folder(os.path.join(root_dir, "test_output"))
         self.replicated_dir = create_folder(
             os.path.join(root_dir, "..", "replicated_dirs")